Refactor parallel training scripts
